+++ src/main/java/com/example/demo/services/MessageService.java
@@ -1,9 +1,14 @@
 package com.example.demo.services;
 
+import java.time.LocalDateTime;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
+
+import com.example.demo.models.Chat;
 import com.example.demo.models.Message;
+import com.example.demo.models.User;
+import com.example.demo.models.role.Role;
 import com.example.demo.repository.MessageRepository;
 
 import lombok.Data;
@@ -13,20 +18,48 @@ import lombok.Data;
 public class MessageService{
 	private final MessageRepository repository;
 	
-	public Message save(Message message) {
-		return repository.save(message);
+	public Message save(Chat chat, User user, Message message) {
+		if(chat.getUsers().contains(user)) {
+			message.setUser(user);
+			message.setChat(chat);
+			message.setTime(LocalDateTime.now());
+			
+			repository.save(message);
+			
+			return message;
+		}
+		
+		return null;
 	}
 
-	public Optional<Message> findById(Long id) {
-		return repository.findById(id);
+	public Message findById(String id) {
+		Optional<Message> message = repository.findById(Long.parseLong(id));
+		if(message.isPresent()) {
+			return message.get();
+		}
+		
+		return null;
 	}
 
-	public Message update(Message message) {
-		return repository.save(message);
+	public Message update(Message message, User user, Chat chat){
+		if(message.getUser().equals(user) && 
+				chat.getMessages().contains(message)) {
+			repository.delete(message);
+			return message;
+		}
+		
+		return null;
 	}
 
-	public void delete(Message message) {
-		repository.delete(message);
+	public Message delete(Message message, User user, Chat chat) {
+		if(chat.getUserRole(user.getId()).equals(Role.ROLE_ADMIN) ||
+				chat.getUserRole(user.getId()).equals(Role.ROLE_MODERATOR)) {
+			repository.delete(message);
+			
+			return message;
+		}
+		
+		return null;
 	}
 
 }
